refactor: foxy withdraw split
