+++ src/features/defi/providers/foxy/components/FoxyManager/Withdraw/WithdrawCommon.ts
@@ -1,0 +1,7 @@
+export enum WithdrawPath {
+  Withdraw = '/',
+  Approve = '/approve',
+  Confirm = '/confirm',
+  ConfirmSettings = '/confirm/settings',
+  Status = '/status',
+}
