+++ src/features/defi/providers/foxy/components/FoxyManager/Withdraw/WithdrawContext.ts
@@ -1,0 +1,10 @@
+import { createContext } from 'react'
+
+import { FoxyWithdrawActions, FoxyWithdrawState } from './WithdrawReducer'
+
+export interface IWithdrawContext {
+  state: FoxyWithdrawState | null
+  dispatch: React.Dispatch<FoxyWithdrawActions> | null
+}
+
+export const WithdrawContext = createContext<IWithdrawContext>({ state: null, dispatch: null })
